+++ server/src/klite/UtilDecorators.kt
@@ -17,6 +17,23 @@ fun RouterConfig.enforceCanonicalHost(host: String) = before { e ->
   if (e.host != host) e.redirect(e.protocol + "://" + host + e.path + e.query, StatusCode.PermanentRedirect)
 }
 
+fun RouterConfig.basicAuth(realm: String = "Auth", authChecker: (name: String, password: Password) -> Boolean) = before { e ->
+  val auth = e.header("Authorization")
+  if (auth?.startsWith("Basic ") == true) {
+    val (user, password) = String(auth.substringAfter("Basic ").base64Decode()).split(':', limit = 2)
+    if (authChecker(user, Password(password))) {
+      e.attr("authUser", user)
+      return@before
+    }
+  }
+  e.header("WWW-Authenticate", "Basic realm=\"$realm\"")
+  throw UnauthorizedException()
+}
+
+fun RouterConfig.basicAuth(users: Map<String, Password>, realm: String = "Auth") = basicAuth(realm) { name, password ->
+  users[name] == password
+}
+
 fun RouterConfig.useHashCodeAsETag() = decorator { e, handler ->
   e.handler()?.also { e.checkETagHashCode(it) }
 }